feat(BR-41b): default API base URL + shared runCli() (Fix 1)
