+++ packages/cowork-desktop/src/cli/index.ts
@@ -1,0 +1,1 @@
+export * from './run.js';
