+++ packages/cowork-desktop/src/config/index.ts
@@ -1,0 +1,1 @@
+export * from './api-base-url.js';
