+++ packages/cowork-desktop/src/index.ts
@@ -22,3 +22,5 @@ export * from './enroll/index.js';
 export * from './registry/index.js';
 export * from './storage/index.js';
 export * from './runner/index.js';
+export * from './config/index.js';
+export * from './cli/index.js';
